Fix incorrect usage of file module
Upstream this was changed to the 'copy' module as I believe it
was originally intended to be used to copy files over. I've
adjusted it so it works more like a permissions configuration
option.

I had to remove the _src variable because when using it with a
list it was causing all sorts of issues since it was expecting
a directory. It was also not necessary
for the use of the file module.